GDB-11259 GDB-11309 GDB-11307
Remove OperationHandler interface
Remove GraphdDBOperator
Remove SinkRecordsProcessor implementation of Operation
Use existing RetryWithToleranceOperator (provided by Kafka) to retry
GraphDB operations (needs testing)
