--- src/main/java/com/ontotext/kafka/operation/OperationHandler.java
@@ -1,7 +1,0 @@
-package com.ontotext.kafka.operation;
-
-import org.apache.kafka.connect.runtime.errors.Operation;
-
-public interface OperationHandler {
-	<E> E execAndHandleError(Operation<E> operation);
-}
